SharePie long に拡張（大きな金額を扱う）

+++ src/main/java/com/example/domain/model/jjugccc2024/advanced/proportion/Amount.java
@@ -3,8 +3,8 @@ package com.example.domain.model.jjugccc2024.advanced.proportion;
 /**
  * *金額
  */
-record Amount(int 金額) {
-    int 差分(Amount 他の金額) {
+record Amount(long 金額) {
+    long 差分(Amount 他の金額) {
         return 金額 - 他の金額.金額;
     }
 }
